Updates, moved testdata out completely
